requirements.txt
# Agentic-Fault-GPT - Requirements

# Python 3.10+ recommended

# -------------------------------
# Core data handling
# -------------------------------
pandas
numpy
pyarrow   # required for parquet

# -------------------------------
# Utilities
# -------------------------------
python-dotenv

# -------------------------------
# Config
# -------------------------------
pyyaml

# -------------------------------
# Streaming (optional)
# -------------------------------
# Only required if you run: python -m simulation.producer --mode kafka
confluent-kafka